+++ src/mol-util/assets.ts
@@ -9,6 +9,7 @@ import { UUID } from './uuid';
 import { iterableToArray } from '../mol-data/util';
 import { ajaxGet, DataType, DataResponse, readFromFile } from './data-source';
 import { Task } from '../mol-task';
+import { File_ as File } from './nodejs-browser-io';
 
 export { AssetManager, Asset };